+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Maintainer: Shiro Takeda
 Change logs of GAMS mode.
 ====================================================
 
+
 Version 6.10
 ----------------------------------
 
@@ -21,6 +22,15 @@ Version 6.10
         (setq gams-close-single-quotation-always nil)
         (add-hook 'gams-mode-hook #'smartparens-mode)
         
+* Changed the default value of `gams-process-command-option` to
+  "logOption=3 pageWidth=100".
+
+* Changed the default value of `gams-system-directory` to "c:/GAMS/37/".
+
+* Changed the default value of `gams-log-file-extension` to "log".
+
+* Updated `gams-setting-sample.el` file.
+
 
 Version 6.9
 ----------------------------------